+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v11.2
 --------------------------
+- Added pdf and csv export for formula analysis
 - Add full usage tab for formula per IFRA category
 - This release may include more changes, for full details please refer to the CHANGELOG